Load files in deterministic order
Before, files would be passed from the Swift extension to the JS script
in a Swift `Dictionary` resulting in a JS object. Since both types are
unordered, the injection of the JS and CSS files were not deterministic.

This commit switches the implementation to use an array instead, which
preserves the order defined by `fileBasenames(forHost:)`.